feat: move e2e test case for edit-history-edit round trip
