+++ e2e/documents/persistence.spec.ts
@@ -5,6 +5,8 @@ import { expect, test } from '../shared/fixtures';
 import {
   openHelloMd,
   openProjectFolder,
+  returnToEditor,
+  selectFirstCommit,
   typeInEditorAndWaitForDebounce,
 } from '../shared/helpers';
 
@@ -60,4 +62,41 @@ test('markdown round-trip: typed content survives a window reload', async ({
   await expect(window.locator('.ProseMirror').locator('h1')).toHaveText(
     'Hello roundtrip'
   );
+});
+
+test('editing after a trip through history keeps every edit', async ({
+  electronApp,
+  window,
+  testProjectDir,
+}) => {
+  await openProjectFolder({ electronApp, window, folderPath: testProjectDir });
+  await openHelloMd({ window });
+
+  await typeInEditorAndWaitForDebounce({
+    window,
+    text: ' hello',
+    waitFor: 500,
+  });
+
+  // Leaving for the history view and back remounts the editor.
+  await selectFirstCommit({ window, commitMessage: 'Set up versioning' });
+  await expect(window.locator('.ProseMirror')).toBeVisible({ timeout: 2_000 });
+  await returnToEditor({ window });
+
+  await expect(window.locator('.ProseMirror').locator('h1')).toHaveText(
+    'Hello hello',
+    { timeout: 2_000 }
+  );
+
+  await typeInEditorAndWaitForDebounce({
+    window,
+    text: ' world',
+    waitFor: 500,
+  });
+
+  const content = fs.readFileSync(
+    path.join(testProjectDir, 'hello.md'),
+    'utf8'
+  );
+  expect(content).toContain('Hello hello world');
 });
